frontend: chats: Move ToolTipIconButton to utils
